chore: Add code `coverage` configuration for Vitest.
It configures Vitest to generate "code coverage reports".

The following changes were made:
  - [x] Add `@vitest/coverage-v8` package as a development dependency.
  - [x] Update `vite.config.js` file to include the `coverage` option.
    This was configured to generate both a terminal summary & detailed
    HTML report.
  - [x] A new `test:coverage` script was added to `package.json` to
    easily run a single test pass with coverage enable.

This change allows developers to measure the effectiveness of the
test suite & identify parts of the codebase that lack test coverage.

P.S.: the base command to run the coverage report is
`npm run test -- --coverage`, this runs in "watch" mode, meaning
it stays active & reruns tests whenever we save a file.